+++ xcube/cli/_gen2/open.py
@@ -19,13 +19,13 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from typing import Sequence, Dict
+from typing import Sequence
 
 from xcube.cli._gen2.genconfig import CubeConfig
 from xcube.cli._gen2.genconfig import InputConfig
-from xcube.core.store import new_data_opener
 from xcube.core.store import DataStorePool
-from xcube.core.store import get_data_store_instance
+from xcube.core.store import get_data_store
+from xcube.core.store import new_data_opener
 from xcube.util.progress import observe_progress
 
 
@@ -38,7 +38,9 @@ def open_cubes(input_configs: Sequence[InputConfig],
         for input_config in input_configs:
             open_params = {}
             if input_config.store_id:
-                opener = get_data_store_instance(input_config.store_id, input_config.store_params, store_pool)
+                opener = get_data_store(input_config.store_id,
+                                        store_params=input_config.store_params,
+                                        store_pool=store_pool)
                 open_params.update(opener_id=input_config.opener_id, **input_config.open_params)
             else:
                 opener = new_data_opener(input_config.opener_id)